Summary is done
All the results are summarized into summary.ser, which has many
columns. It is serialized from a DataFrame. The report can then be
obtained with some one-liners in the example jupyter notebook.

The next stage will focus on the `Conn` module, and `schemes`. The
latter may be organized into another module later.